# Reduce garbage produced by Merkle tree construction