Benchmarks showed that with deeper tree depths the heap and GC pressure is significant during Merkle tree construction time. Besides the size of the trees, there is an avoidable overhead, which is removed by this commit.