@blazember 2018-09-12
The results before and after this change with 4G heap.

Before:
```
Benchmark                                      (depth)  Mode  Cnt     Score     Error  Units
ArrayMerkleTreeBenchmark.createMerkleTree_100        8    ss    5    73.072 ±  96.653  ms/op
ArrayMerkleTreeBenchmark.createMerkleTree_100       10    ss    5    75.293 ±  17.208  ms/op
ArrayMerkleTreeBenchmark.createMerkleTree_100       12    ss    5   110.056 ±  21.788  ms/op
ArrayMerkleTreeBenchmark.createMerkleTree_100       14    ss    5   234.068 ±  18.274  ms/op
ArrayMerkleTreeBenchmark.createMerkleTree_100       16    ss    5   812.853 ±  47.759  ms/op
ArrayMerkleTreeBenchmark.createMerkleTree_271        8    ss    5    78.620 ±  15.860  ms/op
ArrayMerkleTreeBenchmark.createMerkleTree_271       10    ss    5    98.482 ±  14.423  ms/op
ArrayMerkleTreeBenchmark.createMerkleTree_271       12    ss    5   176.803 ±  18.579  ms/op
ArrayMerkleTreeBenchmark.createMerkleTree_271       14    ss    5   517.035 ±  35.367  ms/op
ArrayMerkleTreeBenchmark.createMerkleTree_271       16    ss    5  1684.980 ± 166.694  ms/op
```
NOTE: depth 18 is missing from the before benchmark, since it failed with GC overhead limit exceeded in the 271 case

After:
```
Benchmark                                      (depth)  Mode  Cnt     Score      Error  Units
ArrayMerkleTreeBenchmark.createMerkleTree_100        8    ss    5    54.029 ±   48.413  ms/op
ArrayMerkleTreeBenchmark.createMerkleTree_100       10    ss    5    50.468 ±    8.360  ms/op
ArrayMerkleTreeBenchmark.createMerkleTree_100       12    ss    5    63.583 ±   13.482  ms/op
ArrayMerkleTreeBenchmark.createMerkleTree_100       14    ss    5    90.732 ±   11.928  ms/op
ArrayMerkleTreeBenchmark.createMerkleTree_100       16    ss    5   202.036 ±   13.436  ms/op
ArrayMerkleTreeBenchmark.createMerkleTree_100       18    ss    5   893.611 ±   34.861  ms/op
ArrayMerkleTreeBenchmark.createMerkleTree_271        8    ss    5    47.892 ±    5.394  ms/op
ArrayMerkleTreeBenchmark.createMerkleTree_271       10    ss    5    57.155 ±   10.537  ms/op
ArrayMerkleTreeBenchmark.createMerkleTree_271       12    ss    5    77.816 ±   14.080  ms/op
ArrayMerkleTreeBenchmark.createMerkleTree_271       14    ss    5   151.696 ±   13.682  ms/op
ArrayMerkleTreeBenchmark.createMerkleTree_271       16    ss    5   455.746 ±   10.553  ms/op
ArrayMerkleTreeBenchmark.createMerkleTree_271       18    ss    5  7015.653 ± 2007.379  ms/op
```