@blazember 2018-09-12
Clear is also a time consuming task, but I couldn't find a way for making it faster: (tried [this](https://github.com/blazember/hazelcast/commit/b67fefb2a688db2d17749610856b82ef58790068#r30476489))
```
Benchmark                           (depth)  Mode  Cnt     Score    Error  Units
ArrayMerkleTreeBenchmark.clear_100        8    ss   50    48.262 ± 12.331  ms/op
ArrayMerkleTreeBenchmark.clear_100       10    ss   50    30.228 ±  6.686  ms/op
ArrayMerkleTreeBenchmark.clear_100       12    ss   50    40.731 ±  9.019  ms/op
ArrayMerkleTreeBenchmark.clear_100       14    ss   50    66.860 ± 22.210  ms/op
ArrayMerkleTreeBenchmark.clear_100       16    ss   50   194.983 ± 29.755  ms/op
ArrayMerkleTreeBenchmark.clear_100       18    ss   50   469.235 ± 47.388  ms/op
ArrayMerkleTreeBenchmark.clear_271        8    ss   50    80.678 ± 18.297  ms/op
ArrayMerkleTreeBenchmark.clear_271       10    ss   50    57.283 ±  3.227  ms/op
ArrayMerkleTreeBenchmark.clear_271       12    ss   50    68.062 ±  5.011  ms/op
ArrayMerkleTreeBenchmark.clear_271       14    ss   50   131.458 ±  8.267  ms/op
ArrayMerkleTreeBenchmark.clear_271       16    ss   50   425.160 ± 18.258  ms/op
ArrayMerkleTreeBenchmark.clear_271       18    ss   50  1394.315 ± 89.722  ms/op
```